+++ tests/__init__.py
@@ -1,0 +1,3 @@
+from satella.debug.environment.tracing import enable_environment
+
+enable_environment()